Change structure of default Cat profile pic
Previously, a link was required for a profile pic for each Cat, which
had a default link in the sign up form pointing to the default image.
Now, the default is removed from the form and a profile pic is optional.
The default picture is now loaded through the template if a cat doesn't
have a profile pic saved.

Additionally, the default profile picture is now stored locally, instead
of pointing to an external website where it was hosted.