Consolodate profile edit options into a single page
Otherwise you'd have separate pages for editing email, location, and
timezone. Not great! The downside to this commit is that for
compatibility issues I've had to duplicate baseform.html (rather than
extending it), so any change to the base form or profile form need to be
mirrored. Or I can make the baseform more compatible (currently it
accepts only a single form, whereas the edit profile needs to combine
two forms: one for the User and one for the Profile).